+++ src/BridgeGame.js
@@ -1,7 +1,7 @@
 /**
  * 다리 건너기 게임을 관리하는 클래스
  */
-const InputView = require("./InputView.js");
+const InputView = require("./View/InputView.js");
 const OutputView = require("./OutputView.js");
 
 class BridgeGame {
